perf: Make native methods `@CriticalNative`

+++ packages/react-native-nitro-modules/android/src/main/cpp/core/JArrayBuffer.hpp
@@ -89,8 +89,8 @@ public:
     }
   }
 
-  jlong getBufferSize() {
-    return static_cast<jlong>(_arrayBuffer->size());
+  int getBufferSize() {
+    return static_cast<int>(_arrayBuffer->size());
   }
 
 public:
